Get List of Voices for Specified Region

Obtains a full list of voices for a specific region.

Usage

ms_list_voice(api_key = NULL, region = "westus")

Arguments

api_key

Microsoft Azure Cognitive Services API key

region

Subscription region for API key. For more info, see https://learn.microsoft.com/en-us/azure/cognitive-services/speech-service/regions

Value

A data.frame of the names and their long names.

Details

For more info, see Get a list of voices from the Microsoft documentation.

Examples

if (FALSE) { # interactive()
# List voices for westus
ms_list_voice(region = "westus")

# List voices for eastus
ms_list_voice(region = "eastus")
}